requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.30
pydantic>=2.7.1
pydantic-settings>=2.2.1
playwright>=1.44.0
apscheduler>=3.10.4
python-dotenv>=1.0.1
websockets>=12.0
aiofiles>=23.2.1
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
